requirements.txt
# Google Gemini SDK (new unified SDK - replaces google-generativeai)
google-genai>=1.0.0

# Microsoft Outlook (MSAL + Graph API)
msal>=1.28.0

# HTTP Client
httpx>=0.27.0

# Vector Database
chromadb>=0.4.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Email (Yahoo IMAP/SMTP)
imapclient>=2.3.0

# Dashboard
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.2.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# CLI Interface
rich>=13.0.0
typer>=0.12.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
